% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-build.R
\name{aegis_default_markers}
\alias{aegis_default_markers}
\title{Default marker list used by AEGIS MVP}
\usage{
aegis_default_markers()
}
\value{
A named list of marker genes.
}
\description{
Default marker list used by AEGIS MVP
}
