% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-deconv.R
\name{align_deconv_to_seurat}
\alias{align_deconv_to_seurat}
\title{Align a deconvolution matrix to Seurat spot order}
\usage{
align_deconv_to_seurat(deconv, seu, method_name = "deconv")
}
\arguments{
\item{deconv}{A deconvolution matrix/data.frame with spot row names.}

\item{seu}{A Seurat object.}

\item{method_name}{Method label used in error/warning messages.}
}
\value{
A numeric matrix aligned to \code{colnames(seu)}.
}
\description{
Align a deconvolution matrix to Seurat spot order
}
