% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-methods.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare deconvolution methods}
\usage{
compare_methods(x)
}
\arguments{
\item{x}{An \code{aegis} object.}
}
\value{
Modified \code{aegis} object with \code{x$consensus$comparison} populated.
}
\description{
Computes pairwise method agreement by cell type and spot-level agreement
across methods.
}
