% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aegis.R
\name{is_aegis}
\alias{is_aegis}
\title{AEGIS analysis object}
\usage{
is_aegis(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
\code{aegis} is a lightweight S3 container around a Seurat object and analysis state.
}
