% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aegis.R
\name{is_aegis_multi}
\alias{is_aegis_multi}
\title{Check if object is an AEGIS multi-sample object}
\usage{
is_aegis_multi(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
Check if object is an AEGIS multi-sample object
}
