% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Plot method comparison and consensus outputs}
\usage{
plot_compare(x, type = c("heatmap", "spot_agreement", "consensus_map"))
}
\arguments{
\item{x}{An \code{aegis} object.}

\item{type}{One of \code{heatmap}, \code{spot_agreement}, \code{consensus_map}.}
}
\value{
A ggplot object.
}
\description{
Plot method comparison and consensus outputs
}
