% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-sample.R
\name{split_aegis_by_sample}
\alias{split_aegis_by_sample}
\title{Split an AEGIS object by sample}
\usage{
split_aegis_by_sample(x)
}
\arguments{
\item{x}{An \code{aegis} or \code{aegis_multi} object.}
}
\value{
Named list of per-sample \code{aegis} objects.
}
\description{
Split an AEGIS object by sample
}
