% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-sample.R
\name{summarize_by_sample}
\alias{summarize_by_sample}
\title{Summarize AEGIS results by sample}
\usage{
summarize_by_sample(x)
}
\arguments{
\item{x}{An \code{aegis} or \code{aegis_multi} object.}
}
\value{
A sample-aware summary data.frame.
}
\description{
Summarize AEGIS results by sample
}
