% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-audit.R
\name{plot_audit}
\alias{plot_audit}
\title{Plot audit outputs}
\usage{
plot_audit(
  x,
  type = c("sumdev", "dominance", "entropy", "marker", "smoothness")
)
}
\arguments{
\item{x}{An \code{aegis} object.}

\item{type}{One of \code{sumdev}, \code{dominance}, \code{entropy}, \code{marker}, \code{smoothness}.}
}
\value{
A ggplot object.
}
\description{
Plot audit outputs
}
